Created LFUCache class
The LFUCache uses hashmaps to store:
- File keys and values
- Lists of keys for each frequency
- The frequency of each key

The class also has methods for placing key-value pairs into the cache (put) and retrieving key-value pairs from the cache (get). The frequency of key-value pairs are updated when retrieved from the cache.

I left a block of code in the class file for testing if you need an explanation for how it works. Just uncomment the block and run the file.

I'm not yet sure where to use the get and put functions within the main file code, so I'll need help with figuring that out.